+++ packages/bundlekit-bundler-rollup/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bundlekit/bundler-rollup
 
+## 0.0.11
+
+### Patch Changes
+
+- fe2c801: 修复所有的 bundler 打包问题
+
 ## 0.0.10
 
 ### Patch Changes