+++ packages/bundlekit-bundler-rspack/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bundlekit/bundler-rspack
 
+## 0.0.10
+
+### Patch Changes
+
+- fe2c801: 修复所有的 bundler 打包问题
+
 ## 0.0.9
 
 ### Patch Changes